+++ migrations/20230727154311_accountsBalance.js
@@ -6,9 +6,9 @@ exports.up = function(knex) {
   return knex.schema.createTable('accountsBalance', (table) => {
     table.string('id', 255).unique();
     table.string('userId', 255).notNullable();
-    table.integer('number', 255).notNullable();
+    table.double('number', 255).notNullable();
     table.integer('balance', 255).defaultTo(0);
-    table.integer('date', 255).notNullable();
+    table.double('date', 255).notNullable();
   });
 };
